Initial upload
Initial upload of Django framework powered api. currently it is possible to add new companies by using their ticker symbol, data gets uploaded by using yfinance, it is possible to delete the data from the database and to preview each entry in a form view instead of a table view.